+++ src/GlobalPayments.Api/Entities/Enums/DccProcessor.cs
@@ -1,5 +1,6 @@
 ﻿namespace GlobalPayments.Api.Entities {
     public enum DccProcessor {
+        None = 0,
         Fexco,
         Euroconex
     }