+++ src/GlobalPayments.Api/Entities/Enums/DccRateType.cs
@@ -1,5 +1,6 @@
 ﻿namespace GlobalPayments.Api.Entities {
     public enum DccRateType {
+        None = 0,
         Sale,
         Refund
     }